back/requirements.txt
pillow==12.1.0
numpy==2.4.1
# torch==2.9.1
torchvision==0.24.1
fastapi==0.128.0
uvicorn==0.40.0
python-multipart==0.0.21
segmentation-models-pytorch==0.5.0
